back-end/authentication_services/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
passlib[bcrypt]
python-jose
email-validator
PyJWT
python-dotenv